refactor(svelte2tsx): on: とスプレッド属性を attributes/ へ切り出す
on: ディレクティブ一式を attributes/event_handler.rs、スプレッド属性を
attributes/spread.rs へ純粋移動。上流の EventHandler.ts / Spread.ts に
対応させる。出力・ロジックの変更なし。
