+++ crates/rsvelte_core/src/svelte2tsx/template/attributes/mod.rs
@@ -2,6 +2,8 @@
 
 pub(super) mod action;
 pub(super) mod class_style;
+pub(super) mod event_handler;
 pub(super) mod let_;
+pub(super) mod spread;
 pub(super) mod svg;
 pub(super) mod transition;
